navigation/requirements.txt
actionlib==1.12.1
angles==1.9.12
asn1crypto==0.24.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
backports.functools-lru-cache==1.4
backports.ssl-match-hostname==3.5.0.1
base-local-planner==1.16.7
bondpy==1.8.5
bzr==2.8.0.dev1
cached-property==1.3.1
camera-calibration==1.15.2
camera-calibration-parsers==1.11.13
catkin==0.7.29
catkin-pkg==0.5.2
catkin-pkg-modules==0.5.2
cbor==1.0.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
configobj==5.0.6
configparser==4.0.2
constantly==15.1.0
contextlib2==0.6.0.post1
controller-manager==0.18.4
controller-manager-msgs==0.18.4
cryptography==2.1.4
cv-bridge==1.13.1
cycler==0.10.0
Cython==0.26.1
defusedxml==0.5.0
diagnostic-analysis==1.9.7
diagnostic-common-diagnostics==1.9.7
diagnostic-updater==1.9.7
docker==2.5.1
docker-compose==1.17.1
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
drone-wrapper==1.3.10
dynamic-reconfigure==1.6.5
dynamixel-sdk==3.7.51
empy==3.3.2
enum34==1.1.6
flake8==3.9.2
Flask==1.1.4
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
gazebo-plugins==2.8.7
gazebo-ros==2.8.7
gencpp==0.6.6
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.17
gennodejs==2.0.1
genpy==0.6.16
gevent==22.10.2
greenlet==2.0.2
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
image-geometry==1.13.1
importlib-metadata==2.1.3
incremental==16.10.1
interactive-markers==1.11.5
ipaddress==1.0.17
itsdangerous==1.1.0
Jinja2==2.11.3
joint-state-publisher==1.12.15
jsonschema==2.6.0
kdl-parser-py==1.13.3
keyring==10.6.0
keyrings.alt==3.0
laser-geometry==1.6.7
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lz4==0.10.1
MarkupSafe==1.1.1
matplotlib==2.1.1
mavros==1.16.0
mccabe==0.6.1
mercurial==4.5.3
message-filters==1.14.13
mock==2.0.0
mpi4py==2.0.0
netifaces==0.10.4
nose==1.3.7
numpy==1.13.3
oauth==1.0.1
olefile==0.45.1
PAM==0.4.2
paramiko==2.0.0
pathlib2==2.3.7.post1
pbr==3.1.1
Pillow==5.1.0
psutil==5.4.2
py-trees==0.6.9
py-ubjson==0.8.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.11
pycairo==1.16.2
pycodestyle==2.7.0
pycrypto==2.6.1
pycryptodomex==3.4.7
pydot==1.2.3
pyflakes==2.3.1
Pygments==2.2.0
pygobject==3.26.1
pygraphviz==1.4rc1
PyNaCl==1.1.2
PyOpenGL==3.1.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyserial==3.4
python-apt==1.6.6
python-dateutil==2.6.1
python-gnupg==0.4.1
python-qt-binding==0.4.4
python-snappy==0.5
PyTrie==0.2
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
qrcode==5.3
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.18.4
resource-retriever==1.12.7
roman==2.0.0
rosbag==1.14.13
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosdep==0.22.2
rosdep-modules==0.22.2
rosdistro==0.9.0
rosdistro-modules==0.9.0
rosgraph==1.14.13
rosinstall==0.7.8
rosinstall-generator==0.1.23
roslaunch==1.14.13
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.13
rosmake==1.14.9
rosmaster==1.14.13
rosmsg==1.14.13
rosnode==1.14.13
rosparam==1.14.13
rospkg==1.5.0
rospkg-modules==1.5.0
rospy==1.14.13
rosserial-arduino==0.8.0
rosserial-client==0.8.0
rosserial-python==0.8.0
rosservice==1.14.13
rostest==1.14.13
rostopic==1.14.13
rosunit==1.14.9
roswtf==1.14.13
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.9
rqt-controller-manager==0.18.4
rqt-dep==0.4.9
rqt-drone-teleop==1.3.10
rqt-ez-publisher==0.5.0
rqt-graph==0.4.11
rqt-ground-robot-teleop==1.3.10
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.17
rqt-joint-trajectory-controller==0.17.3
rqt-joint-trajectory-plot==0.0.5
rqt-launch==0.4.8
rqt-launchtree==0.2.0
rqt-logger-level==0.4.8
rqt-moveit==0.5.10
rqt-msg==0.4.8
rqt-multiplot==0.0.10
rqt-nav-view==0.5.7
rqt-paramedit==1.0.1
rqt-play-motion-builder==1.0.2
rqt-plot==0.4.13
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.3
rqt-py-console==0.4.8
rqt-py-trees==0.3.1
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.14
rqt-robot-steering==0.5.10
rqt-rotors==2.2.3
rqt-runtime-monitor==0.5.7
rqt-rviz==0.7.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-virtual-joy==0.1.2
rqt-web==0.4.8
rviz==1.13.30
scandir==1.10.0
SecretStorage==2.3.1
sensor-msgs==1.12.8
service-identity==16.0.0
simplejson==3.13.2
six==1.11.0
smach==2.0.3
smach-ros==2.0.3
smclib==1.8.5
subprocess32==3.2.7
termcolor==1.1.0
terminal-colored-print==0.3.0
texttable==0.9.1
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
topic-tools==1.14.13
trollius==2.0.1
turtlebot3-example==1.2.5
turtlebot3-teleop==1.2.5
Twisted==17.9.0
txaio==2.8.1
typing==3.10.0.0
u-msgpack-python==2.1
Unidecode==1.2.0
unique-id==1.0.6
urdfdom-py==0.4.6
urllib3==1.22
vcstools==0.1.42
wadllib==1.3.2
websocket-client==0.44.0
Werkzeug==1.0.1
wsaccel==0.6.2
wstool==0.1.17
wxPython==3.0.2.0
wxPython-common==3.0.2.0
xacro==1.13.20
zipp==1.2.0
zope.event==4.6
zope.interface==4.3.2
